Expenses form to modal. CategoryModal component
