+++ frontend/src/pages/ExpensesOverviewPage/ExpensesForm.js
@@ -1,7 +1,8 @@
 import { faCirclePlus } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { joiResolver } from '@hookform/resolvers/joi';
-import { Button, DatePicker, Input, Modal, Select } from 'antd';
+import { DatePicker, Input, Modal, Select } from 'antd';
+import CategoryModal from 'components/CategoryModal/CategoryModal';
 import Joi from 'joi';
 import React, { useEffect, useState } from 'react';
 import { Controller, useForm } from "react-hook-form";
@@ -19,12 +20,9 @@ const ExpenseSchema = Joi.object({
   date: Joi.string().required()
 });
 
-const ExpensesForm = ({ isOpen = false, setIsOpen = () => {} }) => {
+const ExpensesForm = ({ isOpen = false, setIsOpen = () => {}, isEditing = false }) => {
   const [ expenseErrorMessage, setExpenseErrorMessage ] = useState('');
-
-  const [openModal, setOpenModal] = useState(false);
-  const [confirmLoading, setConfirmLoading] = useState(false);
-  const [ newCategory, setNewCategory ] = useState('');
+  const [ openModal, setOpenModal ] = useState(false);
 
   const api = useAxios();
   const { register, handleSubmit, control, setValue, reset, formState: { errors } } = useForm({ resolver: joiResolver(ExpenseSchema) });
@@ -32,15 +30,10 @@ const ExpensesForm = ({ isOpen = false, setIsOpen = () => {} }) => {
   const user = useSelector(userSelector.getUser || {});
 
   useEffect(() => {
-    reset();
-  }, [reset])
-
-  useEffect(() => {
-    setNewCategory('');
-    console.log('qweqw')
-  }, [openModal])
-
-  if (!isOpen) return null;
+    if (!isEditing) {
+      reset();
+    }
+  }, [isOpen, isEditing, reset])
 
   const onSubmit = async (data) => {
     try {
@@ -69,59 +62,16 @@ const ExpensesForm = ({ isOpen = false, setIsOpen = () => {} }) => {
     setExpenseErrorMessage('');
   }
 
-  const handleOk = async () => {
-    if (!newCategory) return;
-    setConfirmLoading(true);
-    const data = {
-      categories: [...user.categories, newCategory]
-    };
-    try {
-      const res = await api.put(`/api/user/${user._id}`, data);
-      dispatch(userActions.setUser(res.data));
-      Swal.fire({
-        icon: 'success',
-        title: 'Categories updated',
-        showConfirmButton: false,
-        timer: 1500
-      }).then(() => {
-        setOpenModal(false);
-        setConfirmLoading(false);
-      })
-    } catch(e) {
-      setOpenModal(false);
-      Swal.fire({
-        title: 'Failed to update categories',
-        text: e.message ?? 'Unknown error',
-        icon: 'error',
-        confirmButtonText: 'Close',
-      })
-    }
-  };
-
-  const handleCancel = () => {
-    setOpenModal(false);
-  };
-
   return (
-    <div className='expensesFormWrapper'>
-      <Modal
-        title="Add category"
-        open={openModal}
-        onOk={handleOk}
-        confirmLoading={confirmLoading}
-        onCancel={handleCancel}
-      >
-        <Input
-          className='mb-1'
-          size="large"
-          placeholder='Category name'
-          value={newCategory}
-          onChange={(e) => {
-            setNewCategory(e.target.value);
-          }}
-        />
-      </Modal>
-      <div className='closeForm' onClick={() => setIsOpen(false)}></div>
+    <Modal
+      title="Add expense"
+      open={isOpen}
+      onOk={() => {}}
+      onCancel={() => setIsOpen(false)}
+      className='expensesFormWrapper'
+      footer={[]}
+    >
+      <CategoryModal openModal={openModal} setOpenModal={setOpenModal} />
       <form className='expensesFormCard' onSubmit={handleSubmit(onSubmit)}>
         <Controller
           control={control}
@@ -237,7 +187,7 @@ const ExpensesForm = ({ isOpen = false, setIsOpen = () => {} }) => {
         <input className='primaryButton' type='submit' value='Add expense' />
         {expenseErrorMessage && <p className='errorMessage'>{expenseErrorMessage}</p>}
       </form>
-    </div>
+    </Modal>
   )
 }
