+++ frontend/src/pages/ExpensesOverviewPage/expensesForm.scss
@@ -2,12 +2,6 @@
   display: flex;
   flex-direction: column;
   justify-content: flex-end;
-  position: fixed;
-  left: 0;
-  top: 0;
-  height: 100%;
-  width: 100%;
-  background-color: #000000b5;
   z-index: 1;
 
   .closeForm {
