+++ frontend/src/pages/ExpensesOverviewPage/expensesOverview.scss
@@ -36,7 +36,6 @@
       padding: 15px;
       margin-bottom: 20px;
       border-radius: 24px;
-      min-height: 89px;
 
       .categoryImage {
         max-width: 50px;
